langchain/requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0
mcp>=1.0.0
python-dotenv>=1.0.0
duckduckgo-search>=5.0.0
wikipedia>=1.4.0
yfinance>=0.2.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
requests>=2.31.0
beautifulsoup4>=4.12.0
numpy>=1.24.0
scikit-learn>=1.3.0 